fix: separated tests for choice_only likelihoods and actual tests for sampling from choice-only models
